Update dependencies and fix README.md
Npm reported a security warning on library remark-parse. I've update
remark to version 13, the next would require switching to ES modules.
Please create a new release to get rid of security warnings in projects
that use this script as dev-dependency.

To test functionality, the cli is now run against its own repository
(and it found a minor error :-)